+++ Mu3Library_Base/Samples~/Template/Scripts/Runtime/Sample_MVP/LinkParent/LinkParentView.cs
@@ -1,11 +1,15 @@
 using System;
 using Mu3Library.UI.MVP;
+using UnityEngine;
 
 namespace Mu3Library.Sample.Template.MVP
 {
     public class LinkParentView : View
     {
+        [SerializeField] private RectTransform _childHost;
+
         public event Action OnInstantiateChildButtonClicked;
+        public RectTransform ChildHost => _childHost != null ? _childHost : _rectTransform;
 
 
         protected override void LoadFunc()